app/requirements.txt
attrs==22.1.0
databases==0.6.2
Django==4.1.2
django-cuser==2017.3.16
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
entrypoints==0.4
flake8==5.0.4
frozenlist==1.3.3
gunicorn==19.9.0
httplib2==0.21.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
structlog==22.3.0
uritemplate==4.1.1
urllib3==1.26.12
